Change %d to %ld and cast to long, which is safer than casting 64-bit values to 32-bit int. Fix some int/magma_int_t compatability. (in magmablas)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@9373 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ magmablas/getrf_setup_pivinfo.cu
@@ -149,7 +149,8 @@ adjust_ipiv_batched( magma_int_t **ipiv_array,
     if (offset == 0 ) return;
     if ( m  > MAX_NTHREADS) 
     {
-        fprintf( stderr, "%s: m=%d > %d, not supported\n", __func__, int(m), int(MAX_NTHREADS) );
+        fprintf( stderr, "%s: m=%ld > %ld, not supported\n",
+                 __func__, long(m), long(MAX_NTHREADS) );
         return;
     }
     adjust_ipiv_kernel_batched
@@ -175,7 +176,8 @@ adjust_ipiv( magma_int_t *ipiv,
     if (offset == 0 ) return;
     if ( m  > 1024) 
     {
-        fprintf( stderr, "%s: m=%d > %d, not supported\n", __func__, int(m), int(MAX_NTHREADS) );
+        fprintf( stderr, "%s: m=%ld > %ld, not supported\n",
+                 __func__, long(m), long(MAX_NTHREADS) );
         return;
     }
     adjust_ipiv_kernel
